SEM core functionality added.
